vdr_pi_prefs: Break dependency on vdr_pi.h -- #114
