+++ src/vdr_pi.h
@@ -40,6 +40,7 @@
 #include "vdr_network.h"
 #include "config.h"
 #include "dm_replay_mgr.h"
+#include "commons.h"
 
 // Request default positioning of toolbar tool
 static constexpr int kVdrToolPosition = -1;
@@ -49,60 +50,6 @@ static constexpr int kVdrToolPosition = -1;
 //----------------------------------------------------------------------------------------------------------
 
 class VDRControl;
-
-/**
- * Data storage formats supported by the VDR plugin.
- *
- * Controls how data is structured and stored in VDR files. Each format offers
- * different capabilities for data organization and playback control.
- */
-enum class VdrDataFormat {
-  kRawNmea,  //!< Raw NMEA sentences stored unmodified
-  kCsv,  //!< Structured CSV format with timestamps and message type metadata.
-         // Future formats can be added here
-};
-
-enum class ReplayMode {
-  kNetwork,      // Use network connection
-  kInternalApi,  // Use PushNMEABuffer()
-  kLoopback      // Use WriteCommDriver() on loopback driver
-};
-
-/**
- * Network settings for protocol output.
- */
-struct ConnectionSettings {
-  bool enabled;  //!< Enable network output
-  bool use_tcp;  //!< Use TCP (true) or UDP (false)
-  int port;      //!< Network port number
-
-  ConnectionSettings() : enabled(false), use_tcp(true), port(10111) {}
-};
-
-/**
- * Protocol recording configuration settings.
- *
- * Controls which maritime data protocols are captured during recording.
- * Multiple protocols can be enabled simultaneously.
- */
-struct VdrProtocolSettings {
-  bool nmea0183;                   //!< Enable NMEA 0183 sentence recording
-  bool nmea2000;                   //!< Enable NMEA 2000 PGN message recording
-  bool signalK;                    //!< Enable Signal K data recording
-  ConnectionSettings nmea0183Net;  //!< NMEA 0183 connection settings
-  ConnectionSettings n2kNet;       //!< NMEA 2000 connection settings
-  ConnectionSettings signalkNet;   //!< Signal K connection settings
-
-  ReplayMode replay_mode;
-
-  VdrProtocolSettings()
-      : nmea0183(true),
-        nmea2000(false),
-        signalK(false),
-        replay_mode(ReplayMode::kInternalApi)
-  // nmea0183ReplayMode(NMEA0183ReplayMode::INTERNAL_API)
-  {}
-};
 
 /**
  * Column definition for CSV format files.
